requirements.txt

python-telegram-bot==20.7
aiosqlite>=0.19.0
python-dotenv>=1.0.1
openai>=1.30.0
anthropic>=0.31.0
python-dateutil>=2.9.0.post0
